--- src/components/Post/Post.css
@@ -1,6 +1,0 @@
-.post {
-    width: 200px; /* 원하는 너비로 설정 */
-    border: 1px solid black; /* 테두리 설정 */
-    padding: 10px; /* 내부 패딩 설정 */
-    margin: 10px; /* 주변 마진 설정 */
-  }
