+++ src/components/post/PostStyles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const PostStyles = styled.div`
+width: 200px; 
+border: 1px solid black;
+padding: 10px; 
+margin: 10px; 
+`;
